migrations: skip repos that cannot be fetched
On systems where users do not allow everybody to fetch their repo, you need to
be able skip certain repos.

Currently, if you cannot authorize yourself to read a migrated repo, abapGit
exits.

With this patch applied, failed authorization attempts are notified to the user
and the repo is skipped.

Signed-off-by: Jakub Filak <jakub.filak@sap.com>