html_utils: URL unescape parameter fields
For some strange reason even post data is URL escaped on our
servers.

How I discovered this issue:
1. go to stage
2. leave some files in the state ?
3. click the Commit link
4. get the error 'process_stage_list: unknown method %3F'

URL decode '%3F' and see that the value is '?'.

The value is set to the hidden input fields in its unescaped form
in the function submitSapeventForm() but the value is escaped in
the method lif_gui_page~on_event of the class lcl_gui_page_stage.