dialog for overwriting local objects rewritten
When Pulling local objects are checked if they have been modified. If so there was a confirmation popup for each objects. This has been replaced with a popup table allowing to have a look at all modified objects at once and selecting a checkbox to confirm that all selected objects should be overwritten.