Function Group Include Handling
- Fixed the handling of function group includes, since abapGit tried to judge the origin of an include by its creator and excluded includes that had SAP as creator. Unfortunately, creator is also set to SAP if objects originate from a system in which they have been exported in an AAK package, since SAP anonymizes objects on export (i.e. sets creator to SAP). Since REPOSRC can not be modified in the system, it is not possible to fix the faulty creator for the includes.
The implemented fix is such that instead of checking the creator, the system checks the status of the namespace in the system for includes with a namespace and accepts any include with a namespace that is set to 'producer' in the system. For objects without a namespace currently the creator is still checked.