Fix issue 1298 - 2FA behind HTTP Proxy
Fix for issue 1298. Previously code was not checking if HTTP proxy authenticaiton was required. Pushed basic method of getting a HTTP client with proxy configured in to 2FA base class. There's some overlap with logic in zcl_abapgit_http%3D>create_by_url but not extracted method at this point