Mass activation: re-check DDIC objs in transport
DDLS objects are not cheked in a transport in their deserialize method.
The parallel activation does not call RS_WORKING_OBJECTS_ACTIVATE which
checks in the activated object in a transport.
Therefor, DDLS objecst are never checked in a transport when
the parallel activation is in place (the option experimental features
is enabled - which is enabled by default on all our systems).

I wanted to use TRINT_OBJECTS_CHECK_AND_INSERT because this function
is used by ADT:

CL_WB_ADT_REST_RESOURCE~post
    CL_WB_ADT_REST_RESOURCE_DATA
    ~if_wb_adt_rest_resource_data~get_transport_key
        CL_WB_OBJECT~get_transport_key -> PGMID%3DR3TR, OBJTYPE%3DDDLS
    CL_WB_ADT_REST_RESOURCE~lock
        CL_WB_ADT_REST_RESOURCE~check_obj_in_transport_request
            CL_ADT_CTS_MANAGEMENT~check_objects_in_request
                CL_ADT_CTS_MANAGEMENT~LCL_DB_PERSISTENCE~check_objects
                    CALL FUNCTION CTS_WBO_API_CHECK_OBJECTS
                        CALL FUNCTION TRINT_WBO_IMPL_CHECK_OBJECTS
                            SAPLCTS_WBO_API_IMPL /
                            LCTS_WBO_API_IMPLF06 /
      FORM CALL_OBJECTS_CHECK_AND_INSERT
                                CALL FUNCTION
                                TRINT_OBJECTS_CHECK_AND_INSERT


But it is not enough to call this function, hence, I use RS_CORR_INSERT
which was mentioned by @larshp in our slack conversation. The function
internally use calls TRINT_OBJECTS_CHECK_AND_INSERT.

I tested this patch with modification of a CDS view and also with 
adding a new CDS view.

Closes #1466