tp: use the correct transport type in the tr popup
We are developing in an infinity system which is somehow cloned into a
new releases system. When we need to backport a bug fixes we do
cherry-pick the commits and pull them the release system. However, in
those systems we need to use the task type Repair.

Hence, we need to overwrite the default value of the parameter
wi_task_type of the function TRINT_ORDER_CHOICE which is 'S'
(Development/Correction) with the value 'R'.

I use the function TRINT_GET_REQUEST_TYPE and for the sake of simplicity
I assume that it is enough to know the transport type of the repository
package - in other words, it is no necessary to check every deserialized
object.

Related to #1446