VCLS jump: open view cluster in new window
Before this commit VCLS jump calls VIEWCLUSTER_MAINTENANCE_CALL
which opens inplace. But that's not compatible with abapGit UI
and nothing is visible.

And the semantics changed. The above mentioned FM call the view
cluster for data maintenance and not the development object.
To be aligned with abapGit logic now the transaction SE54 is called.