AVAS: report detailed errors
Unfortunately, we are versioning AVAS objects that can be instantiated
only in SAP systems but we use the repo also in sandbox Customer
systems.

Whenever I open the repository I get an exception of the type
cx_pak_invalid_data.

Without this commit I see the error message:
"AVAS error".

With this commit I see the error message:
"AVAS $GUID: invalid data: The functionality is not supported in
customer systems"