SICF: Delete data from  ICFAPPLCUST
The data in table ICFAPPLCUST (Application Customizing Data) isn't
deleted by cl_icf_tree=>if_icf_tree~delete_node( ). Before this commit
is applied it's not possible to delete ICF nodes with data in that table.
With this commit we delete the data in the table. It's the same as
transaction SICF does.