objects: replace rs_c_true with abap_true
The constant rs_c_true comes from the package RS which is no available
in pure BASIS systems.

I believe that the constant rs_c_true was used by mistake.
The constant was assigned to a variable which is of the type abap_bool
and the variable is initialized with abap_false.

Related to #1826