Sapscript : sequence of Windows in tdlines files
Hello Lars,

I have moved the code to a new private method _sort_tdlines_by_windows
But I don't add 'ASSERT sy-subrc = 0.'. 
The reason is that during my tests I found some windows in the form without elements inside (windows is not used anymore but it sill exists. The devlopper ddin't remove it). In this case the result of the ASSERT is a dump.
Using the loop from SY-TABIX there is no risk. If no entry exists in the internal table then the loop is not executed.

if needed I can add the following condition before doing the loop :
       IF line_exists( lt_lines[ tdformat = lv_elt_windows
                                tdline = ls_form_windows-tdwindow ] ).
Christophe