Sapscript : sequence of Windows in tdlines files (#2191)
* Sapscript : sequence of Windows in tdlines files

Hello Lars,

You can find a code modification in the method _read_form of class zcl_abapgit_object_form related to the issue #2188
I have added a logic to sort by windows name the contains of the internal table ET_LINES . It allows to have always thre same sequence of  windows code in the file (xxxxxxxxxxgit.form.tdlines). Then file comparison will work with all export.

Regards
Christophe

* Sapscript : sequence of Windows in tdlines files

Hello Lars,

I have moved the code to a new private method _sort_tdlines_by_windows
But I don't add 'ASSERT sy-subrc = 0.'. 
The reason is that during my tests I found some windows in the form without elements inside (windows is not used anymore but it sill exists. The devlopper ddin't remove it). In this case the result of the ASSERT is a dump.
Using the loop from SY-TABIX there is no risk. If no entry exists in the internal table then the loop is not executed.

if needed I can add the following condition before doing the loop :
       IF line_exists( lt_lines[ tdformat = lv_elt_windows
                                tdline = ls_form_windows-tdwindow ] ).
Christophe

* Sapscript : sequence of Windows in tdlines files

I have done the adjustments for the CLEAR and to be  compatible with 702

Christophe

* Sapscript : sequence of Windows in tdlines files

Whitespace at end of line removed