repo: add folder logic and ign. sub.pkg to the new repo popup
This patch adds the parameters Folder logic and Ignore subpackages to
the new online repo popup to help us deal with old package which do not
follow the prefix approach for subpackages and to allow us to version
superpackages (which are shared by many developers) in separate
repositories.

I am not sure how to add a check-box to the popup window. I think I need
to find the right table with the righ column of the type char(1).

Closes #2395