objects: add a method checking presence of a file
The method is called 'contains' because then you can write:

  if mo_files->contains( iv_ext = 'abap' ) = abap_true.

In newer system, you can even leave out the suffix '= abap_true'.

I am not sure how to make the logic deciding how to the select
shared with the method read_file.

For future developers: if you change the type of raised exception of the
method read_file when the passed file is not found, you can stop using
the method contains and you can save some time in the cases where it is
highly unexpected that the file is missing. You could also remove the
method contains which will have the benefit of having one place with the
select statement.