RS_CUA_INTERNAL_WRITE another fix (#2656)
* Update zcl_abapgit_objects_program.clas.abap

Found a situation on my project where on pulling code from a zip file, the error RS_CUA_INTERNAL_WRITE sy-subrc = 2 popped.
The error was due to the content of XML file exported : <CUA><ADM>ACTCODE was containing an 'X' value for a single file. I corrected the file putting it blank and it worked. But it's a fastidious task so I fixed the method ZCL_ABAPGIT_OBJECTS_PROGRAM->auto_correct_cua_adm in order to fix this field also when it contains only an 'X'. Tested ok on my system so I propose this small change.

* Update zcl_abapgit_objects_program.clas.abap

* Update zcl_abapgit_objects_program.clas.abap

Use the same check SAP does in form check_adm of include LSMPIF03 for the CUA values.