oo: fix dump during deserialization
A class with redefintions was causing dumps.

I have learned that SAPLSEOR:INSERT_REDEFINITIONS tries to insert all redefinitions
without checking if they already exist.

To enable the check, we have to make sure wbinactive is X.

So, we have to add CALL FUNCTION 'SEO_CLIF_INACTIVE_CHECK' which sets
wbinactive to X.

But we have to also set suppress_access_permission because otherwise
the function seo_clif_access_permission set wbinactive  to ' '.

Finally we have to pass state %3D 'I' otherewise wbinactive is again set to ' '.

I haven't changed the state in the old release branch because I cannot
test this change.