otgr: initial object support
This commit replaces the generic SOBJ support with native support for
the object type OTGR and the format is not backward compatible,
so repos containing SOBJ serialized objects could not be pulled anymore
(it's necessary to pull them with previous version of abapGit and push
them with the new version of abapGit).

The new OTGR handler supports only master language description because
there is no API which would allow us to maintain other languages and
I am not willing to re-implement parts of ABAP platform just to version
texts which nobody cares about. Also, the texts dealing part is not
that straightforward - one has to think about keeping the texts
in sync which requires lot of code.

I did the following optimizations to the XML format:
- removed the activation state from all elements because we
  always serialize only Active and deserialize as Inactive
- removed the foreign key name from texts and other details
  because we know which object type group is currently being handled
- used short names for root elements of details to avoid redundancy
  because the detail's table name is used for child XML elements -
  e.g. TEXTS vs. CLS_TYPE_GROUPT.

Closes #1515