Set filter for file open and save dialogs / Add new setting for default commit comment (#3371)
* Set filter for file open and save dialogs

Set the file filter dropdown and default to *.ZIP in file open and save dialogs

* Add new setting for default commit comment

New abapGit setting to define a default for the comment used for commits. Object and/or filename can added to the comment automatically, which works similar to GitHub if a single file is changed. If multiple objects are changed, then object/file count is included in the default for the comment.

* Update CLAS ZCL_ABAPGIT_GUI_PAGE_COMMIT

* Fix text for $object

Fixed case when one object with multiple files was changed

* Pretty print

Co-authored-by: Lars Hvam <larshp@hotmail.com>