Fix for Java GUI so new repo can be defined (#4655)
* Fix for Java GUI so new repo can be defined

The HTML demo programs are able to pass values from a form, so why does abapGit not work with Java GUI?
I've been trying to figure this out. The key to a break-through was not any fancy code analysis, it was "brute force".
I added a method in ZCL_ABAPGIT_HTML_FORM to replace the generated HTML by a copy of the resulting HTML. Then I removed things until POST-data was filled, and added things back until it broke again.

The main issue was of course the complete lack of POST-data for the event.
This is solved by adding an action attribute to the form element and removing the form-action attribute from the 'default' submit button.

The biggest surprise was that radion buttons don't work as they should. Values are returned as if both radio buttons (in the new online repository form) are checked - which of course is not possible. I tried enabling the query table in the HTML viewer just in case this was specific to the POST-data, but no such luck, it is probably a problem with the HTML viewer itself. You won't find any samples from SAP with radio buttons either.
The bug is possibly fixed by the patch in Note 2832978.

The solution is far from perfect, but at least it is now possible to add new repositories in Java GUI.
The most annoying problem that remains is that if a dialog is opened to select a package or branch the existing input data are lost. One step at a time though...

* Update src/ui/zcl_abapgit_html_form.clas.abap

Linter fix

Co-authored-by: abaplint[bot] <24845621+abaplint[bot]@users.noreply.github.com>

* Single-parameter calls on one line

One extended line, mind you...

* One argument per line - when possible

Don't break the string expression up over more lines thna neessary.

* Keep line length under 120 characters

* Playing catch-up

Fixing the conflict after another pull-request was merged

* Catch up with main branch

I didn't understand from the conflict resolution view on github the total difference between main and my fork. Looked at the source file and it became clear. 
So hopefully this is the last commit... (and hopefully that won't be "famous last words")

* Time to learn about rebase...

Removed duplicate statement to get number of rows

* Conflicting line reset

Reset to main
<<<<<<< kjetil-kilhavn/javagui-fix
          ii_html->add( |<td><input type="text" name="{ lv_cell_id }" id="{ lv_cell_id }"|
                     %26%26 | value="{ lv_value }"{ lv_readonly }></td>| ).
=======
          ii_html->add( |<td><input type="text" name="{ lv_cell_id }" id="{
                        lv_cell_id }" value="{ lv_value }"{ lv_readonly }></td>| ).
>>>>>>> main

Now, may I have all green lights please?

Co-authored-by: abaplint[bot] <24845621+abaplint[bot]@users.noreply.github.com>
Co-authored-by: Lars Hvam <larshp@hotmail.com>