UI: Sticky Top (#5061)
* UI: Sticky Top

With this change, the top bar will become sticky. Meaning, it remains on the top of the page when scrolling down.

It's CSS + JS since IE does not support the "sticky" CSS feature.

There's a bit of a "wiggle" when scrolling begins and alignment is not the best on the diff page (which has full width). Maybe some of you CSS wizards can figure these things out.

Closes #5035

* Upload with LF + UTF8

* Remove todo

- already includes changes of 5083

Co-authored-by: Christian Günter <christianguenter@googlemail.com>
Co-authored-by: Lars Hvam <larshp@hotmail.com>