Add user exit for adjusting display of filenames (#5185)
The filenames can become very long, for example, usage of many sub-packages, namespaces, objects like function groups, or BSP/Fiori apps with many sub-objects, or a long paths as starting folder. This can make the repository view very busy or render the output as a very wide table (see #5169, #5178).

This exit allows to adjust the filename according to your own rules. 

Here's an example for shortening the display of files for sub-objects:

```abap
  METHOD zif_abapgit_exit~adjust_display_filename.

    DATA:
      lv_path TYPE string,
      lv_name TYPE string,
      lv_ext1 TYPE string,
      lv_ext2 TYPE string.

    SPLIT iv_filename AT '.' INTO lv_path lv_name lv_ext1 lv_ext2.

    IF lv_ext2 IS INITIAL.
      " Main object
      rv_filename = iv_filename.
    ELSE.
      " Sub object
      rv_filename = |…{ lv_ext1 }.{ lv_ext2 }|.
    ENDIF.

  ENDMETHOD.
```