SOTR: Fix handling of alias names
When deserializing a Web Dynpro that contains OTR texts using aliases, abapGit raised error "Invalid alias" if the target package was different from the original package. The reason was that the alias name contained the package name as well.

The new serialize removes the old package name from the alias and deserialize sets it to the new package.The logic is backward compatible.

Closes #5232