"New Online Repo": Derive package/name from URL
Adds a button in the "New Online Repo" dialog to derive the SAP package from the URL. 

The repo name part of the URL will be used as the package name. If it does not start with Y/Z, it adds a `$` in front. 

If the SAP package already exists, then its short text will be used as the display name.