TABL: Fix deserialize if foreign key exists
In case a foreign key definition already exists for a table and is referenced in another object - like a maintenance view - the activation would fail with error "deleted foreign key is still used":


The reason was that the first deserialize phase activated the table without any foreign keys (to prevent dependency issues, in case the check table doesn't exist). Now only non-existing foreign key definitions will be removed in the frist phase avoiding the error.

Closes #5694