CLAS,ENHO: Fix inactive enhancements
When pulling a change to an existing class, existing enhancement where inactivated and the enhancement code deleted. 

This was caused by unnecessarily calling `CALL FUNCTION 'SEO_CLASS_CREATE_COMPLETE'` in `zcl_abapgit_oo_class->create`, twice.

The change avoids the second call if the class properties and attributes have *not* been changed. Interfaces are handled the same way. 

Tested with CI tests for classes and interfaces:


Closes #5809