Prevent pull in case of duplicate files
If there are multiple files with the same filename, pulling will raise an error. To allow pulling requires cleaning up the repository. Keep on of the files and remove the others from the repository.

Closes https://github.com/abapGit/abapGit/issues/4393

PS: It's unlikely that this situation happens when using exclusively abapGit to stage the repository since checks are in place to prevent such. However, it is possible if one makes changes to the repository using other tools like GitHub or VSCode.