Pull enabled for customizing tables
This commit enables you to download table-content to your system.
For safety reasons, this is limited to customizing tables (DD09L-TABART = 'APPL2') in custom tables (DD09L-AS4USER <> 'SAP'). Only those tables will be written to the system, that has been selected by the user. If the system has been set to change recording in SCC4, the user will be prompted for a customizing request, where the entries will be added to.

What is missing:
> The solution does not fit perfectly in the existing architecture. I am thinking of moving the whole persistance logic to another place outside the deserializer.
> Update of existing table entries is not working, only INSERT/DELETE is allowed.