Enhanced Serializer Interface
As proposed during community call and in https://github.com/abapGit/abapGit/issues/4187, here are the enhancements of the serialized interface.

https://larshp.github.io/abapGitCC20230404/#/10

New Methods

- get_deserialize_order 
- map_object_to_filename
- map_filename_to_object

New Parameters

- changed_by, jump (iv_extra, iv_ext)

Note: The three new methods are not implemented or called yet. This will happen in a subsequent PR. Until then, a syntax check of serializer classes will show warnings which can be ignored.