SICF: Remove special TADIR logic
Leaveraging the new serializer interface. The change removes the special logic translating TADIR into ITEM for `SICF` and replaces it with an equivalent mapping between object and filenames. 

The implementation of the serializer class is now much more straight forward. It's always using `ms_item` to refer to the object which is equivalent to the tadir entry like for other objects. 

Ref #3453, #4187, #5064 
Closes #6172