Handling unsupported objects in deserialize proces
With this change, unsupported objects are shown with a different icon and "Object type not supported" message in the pull popup. I'm not aware of any option to disable a checkbox for a ALV row. So it's still possible to select the objects and try to pull them, which will inevitably result in an error. Alternatively, we could remove those entries from the popup completely. Let me know what you prefer.

Also raises an exception in AFF constructor if object type is not supported (as required in [docs](https://docs.abapgit.org/development-guide/serializers/serializers.html#constructor))

Ref https://github.com/abapGit/abapGit/issues/4898#issuecomment-1601647040