Do not assign T packages to a transport request
When doing pull to a new T package, then abapgit asks for a transport request.
T packages are local and they do not need to be assigned to a transport
request.

This change updates the condition which decides if a new package should be
assigned to a transport request or not.