Fix deletion/uninstall when using local settings
If the local options "ignore subpackages" or "only local objects" are turned on, then deleting all objects or uninstalling the repository will also delete all subpackages and objects included in the subpackages and objects that are not local to the system.

The fix centralizes the tadir object selection and ensures that local settings "ignore subpackages" or "only local objects" are respected when selecting objects.