New feature: Original system
It's the longest outstanding issue from about 7 years ago and here's the solution to it. 
Closes #641

You can now maintain the original system in your repo settings. This setting is then included in `.abapgit.xml`. When deploying such repository into a downstream system like Test or QA, abapGit with perform two addition steps:

1. Check the original system of existing objects

If an object has a different original system than the one in the repository, abapGit will not allow overwritting the object.

2. Update the original system after deserializing objects

Objects are set to the original system of the source. This will prevent changes to these objects.

Happy testing and Merry Christmas!